refactor dialog utilities of the "use prompt command"
